backend improvement en de groep settings pagia layout gewijziged

+++ server/api/moments/[group_id]/[image_id]/index.delete.ts
@@ -1,9 +1,15 @@
 export default defineSupabaseEventHandler(async (event, user, client) => {
 
-    const { image_id } = getRouterParams(event)
+    if (!user) return useReturnResponse(event, unauthorizedError);
 
+    const { image_id } = getRouterParams(event)
     const { error } = await client.from('posts').delete().eq('id', image_id)
+    
     if (error) return useReturnResponse(event, internalServerError);
+
+    /*
+    ************************************************************************************
+    */
 
     return useReturnResponse(event, {
         status: {
